Fixed entity mapped Added controller Added angular folder Added custom json deserializer

+++ src/main/java/org/abpc/repository/ClassesRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.repository.CrudRepository;
 import javax.transaction.Transactional;
 
 @Transactional
-public interface ClassesRepository extends CrudRepository<Classes, Long> {
+public interface ClassesRepository extends CrudRepository<Classes, Integer> {
 }
