+++ src/main/java/org/abpc/repository/MedalRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.repository.CrudRepository;
 import javax.transaction.Transactional;
 
 @Transactional
-public interface MedalRepository extends CrudRepository<Medal, Long> {
+public interface MedalRepository extends CrudRepository<Medal, Integer> {
 }
